pyproject.toml
[tool.poetry]
name = "ml-service"
version = "0.1.0"
description = "Ml service project"
authors = ["Trubin Daniil"]
readme = "README.md"
packages = [{ include = "models", from = "src" },
            { include = "service", from = "src" },]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
pandas = "^2.1.3"
scikit-learn = "^1.3.2"
uvicorn = "^0.24.0.post1"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.23"
aiosqlite = "^0.19.0"
python-jose = "^3.3.0"
celery = "^5.3.6"
flower = "^2.0.1"
redis = "^5.0.1"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
plotly = "^5.18.0"
nbformat = "^5.9.2"
seaborn = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

